requirements.txt
streamlit==1.31.0
python-dotenv==1.0.0
langgraph==0.2.0
langchain-core==0.3.0
langchain-google-genai==2.0.0
langchain-community==0.3.0
langchain-huggingface==0.1.0
pypdf==4.0.0
docx2txt==0.8
beautifulsoup4==4.12.3
lxml==5.1.0
pydantic==2.6.0
google-generativeai==0.8.0